+++ src/lib/models/types.ts
@@ -21,8 +21,10 @@ export abstract class Type {
     /**
      * Visit this type, returning the value returned by the visitor.
      */
-    visit<T>(visitor: TypeVisitor<T>): T {
-        return visitor[this.type](this as never);
+    visit<T>(visitor: TypeVisitor<T>): T;
+    visit<T>(visitor: Partial<TypeVisitor<T>>): T | undefined;
+    visit(visitor: Partial<TypeVisitor<unknown>>): unknown {
+        return visitor[this.type]?.(this as never);
     }
 }